Move in memory storage into its own package/dir
