+++ k8s/service_controller/server/mem/in_mem_service_storage_test.go
@@ -1,4 +1,4 @@
-package server
+package mem
 
 import (
 	"strings"
